+++ google2/CHANGELOG.md
@@ -10,6 +10,7 @@ Changed:
 - Make `genDiskName` non-empty
 - Bump `http4s` version to `0.21.5`, `scalatest` to `3.1.2`, `fs2-io` to `2.4.2`
 - Add `autoDeleteDiskDeviceName: Set[DeviceName]` to `deleteInstance` method
+- get nodepool returns an option
 
 Added:
 - Add `GoogleDiskService` and `GoogleDiskInterpreter`
@@ -19,7 +20,7 @@ Added:
 - Add Generator for `DiskName`
 - Add Kubernetes client APIs for creating service accounts, roles and role bindings
 
-SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google2" % "0.10-5df3b81"`
+SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google2" % "0.10-TRAVIS-REPLACE-ME"`
 
 ## 0.9
 Changed: